// src/api/model/itemTodo.ts
/**
 * Expeneses API
 * API providing data for expeneses new project|
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: tomerk@kkl.org.il
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * task
 */
export interface ItemTodo { 
    /**
     * unique id of the task
     */
    id?: number;
    /**
     * task description
     */
    description?: string | null;
    done?: boolean;
    /**
     * unique id of the user
     */
    userId?: number;
    fullName?: string;
}