// src/api/model/users.ts
/**
 * Expeneses API
 * API providing data for expeneses new project|
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: tomerk@kkl.org.il
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * the otp responses
 */
export interface Users { 
    /**
     * unique id of the user
     */
    id?: number;
    fullName?: string;
}